+++ CHANGELOG.md
@@ -17,3 +17,10 @@
 - Removed some debug messages from the console
 - Fixed a bug in the Command Binder that was causing speech marks to be included in string tokens after binding
 - Included a copy of UnityDefaultThemes in the package so that the console would render correctly once imported to another project
+
+## [1.2.1] - 20/02/2025
+## Console feature update
+- Fixed some layout issues with the Console
+- Added a scroll bar which appears when the console exceeds a maximum size
+- Made it so the console correctly scrolls all the way to the bottom when a new message is added
+- Added a command history to the Console. Using the up and down arrow keys allows scrolling through history to quickly copy commands